move stack config to config file
